requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-appconf==1.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-imagekit==4.1.0
djangorestframework==3.14.0
pilkit==2.0
Pillow==9.2.0
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
